Initial
This function runs stepwise model selection and outputs a set of
statistics.
Specifically designed for an F2 eQTL experiment.
Main feature is that it allows the user to input a starting position
for the QTL search, thus not penalizing the first QTL included if it is
where the user a priori expected (e.g. cis eQTL)